Fix Fedora 32 gcc/clang warnings
Component: AOSP, OpenGL, Vulkan

VK-GL-CTS issue: 2350

Affects:
KHR*direct_state_access.vertex_arrays*
KHR*core.arrays_of_arrays.sized_declarations_invalid*
dEQP-VK.geometry.*

+++ external/vulkancts/modules/vulkan/conditional_rendering/vktConditionalDrawTests.cpp
@@ -456,7 +456,7 @@ tcu::TestStatus ConditionalDraw::iterate (void)
 	submitCommandsAndWait(m_vk, device, queue, m_cmdBuffer.get());
 
 	// Validation
-	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5 + WIDTH), (int)(0.5 + HEIGHT));
+	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5f + static_cast<float>(WIDTH)), (int)(0.5f + static_cast<float>(HEIGHT)));
 								  referenceFrame.allocLevel(0);
 
 	const deInt32 frameWidth	= referenceFrame.getWidth();
